kubecli: gate generated mock behind kubevirt_mock build tag
Downstream consumers on k8s.io/client-go v0.35+ cannot compile
kubecli because generated_mock_kubevirt.go imports the removed
storagemigration/v1alpha1 package. The mock is only needed for
KubeVirt unit tests, so gate it (and the mock test helpers) behind
the kubevirt_mock build tag.

Fixes #18382
