+++ .bazelrc
@@ -27,6 +27,8 @@ run:crossbuild-s390x  --incompatible_enable_cc_toolchain_resolution --platforms=
 test:crossbuild-s390x --platforms=@io_bazel_rules_go//go/toolchain:linux_s390x_cgo --host_javabase=@local_jdk//:jdk
 
 build --define gotags=selinux
+test --define gotags=selinux,kubevirt_mock
+coverage --define gotags=selinux,kubevirt_mock
 
 # CentOS Stream version selection (default to CS9)
 build --define=centos_stream_version=9
